Moving all security config to one class
